backend/MindMapGeneration/requirements.txt
annotated-types
anyio
certifi
cffi
charset-normalizer
click
cryptography
exceptiongroup
fastapi
filelock
fsspec
h11
hf-xet
huggingface-hub
idna
Jinja2
joblib
MarkupSafe
mpmath
networkx
numpy
packaging
pdfminer.six
pdfplumber
pillow
pycparser
pydantic
pydantic_core
pypdfium2
python-multipart
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
sniffio
starlette
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn
python-dotenv
